refactor: extract CCTV media and radio directory providers
