+++ docs/CURRENT-STATE.md
@@ -1,5 +1,15 @@
 # God's Eye View Current State
 
+## CCTV and radio provider modules
+
+CCTV catalog acquisition, source normalization and frame/media delivery now live
+in separate modules. Each CCTV factory owns its catalog and health state; its
+`sourceRoot` option resolves relative source files against the application root.
+Radio Browser station normalization, restricted outbound transport and directory
+caching are separate modules. Node-only package entries expose both provider
+factories. Routes, payloads, fallback behavior and existing dev/preview hook
+registration are preserved; browser rendering is unchanged.
+
 ## Terrain, traffic, fire and bike-share provider modules
 
 Local composition now imports separate Node modules for Re:Earth heights,
